rss feed should also get sorting
