+++ dev/Sonos-Kids-Controller-master/src/app/rssfeed.service.ts
@@ -1,11 +1,13 @@
-import { Injectable } from '@angular/core';
-import { Observable, defer, throwError, of, range } from 'rxjs';
-import { retryWhen, flatMap, tap, delay, take, map, mergeMap, mergeAll, toArray, switchMap } from 'rxjs/operators';
-import { environment } from 'src/environments/environment';
+import { ExtraDataMedia, Utils } from './utils';
+import { Observable, defer, of, range, throwError } from 'rxjs';
+import { delay, flatMap, map, mergeAll, mergeMap, retryWhen, switchMap, take, tap, toArray } from 'rxjs/operators';
+
 import { HttpClient } from '@angular/common/http';
+import { Injectable } from '@angular/core';
+import { Media } from './media';
 //import { xml2json } from 'xml-js';
 import { RssFeed } from './rssfeed';
-import { Media } from './media';
+import { environment } from 'src/environments/environment';
 
 declare const require: any;
 const xml2js = require('xml2js');
@@ -25,11 +27,7 @@ export class RssFeedService {
     id: string,
     category: string,
     index: number,
-    shuffle: boolean, 
-    aPartOfAll: boolean,
-    aPartOfAllMin: number,
-    aPartOfAllMax: number,
-    manualArtistcover: string
+    extraDataSource: ExtraDataMedia
   ): Observable<Media[]> {
     this.url = 'http://' + ip + ':8200/api/rssfeed?url=' + id;
     return this.http.get(this.url/*, { responseType: 'text' }*/).pipe(
@@ -47,23 +45,8 @@ export class RssFeedService {
             type: 'rss',
             category,
             index,
-          };
-          if (manualArtistcover) {
-            media.artistcover = manualArtistcover;
           }
-          if (shuffle) {
-            media.shuffle = shuffle;
-          }
-          if (aPartOfAll) {
-            media.aPartOfAll = aPartOfAll;
-          }
-          if (aPartOfAllMin) {
-            media.aPartOfAllMin = aPartOfAllMin;
-          }
-          if (aPartOfAllMax) {
-            media.aPartOfAllMax = aPartOfAllMax;
-          }
-          console.log(media);
+          Utils.copyExtraMediaData(extraDataSource, media)
           return media;
         });
       }),
@@ -71,54 +54,4 @@ export class RssFeedService {
       toArray()
     );
   }
-
-  //async parseXmlToJsonRss(xml) {
-    // With parser
-    /* const parser = new xml2js.Parser({ explicitArray: false });
-    parser
-      .parseStringPromise(xml)
-      .then(function(result) {
-        console.log(result);
-        console.log("Done");
-      })
-      .catch(function(err) {
-        // Failed
-      }); */
-
-    // Without parser
-    //return await xml2js
-    //  .parseStringPromise(xml, { explicitArray: false })
-    //  .then((response) => response);
-  //}
-  
-  
-  
-  
-  
-  
-  // test (){
-  //   var url='https://www.antennebrandenburg.de/programm/hoeren/podcasts/Zappelduster_Podcast/podcast.xml/feed=podcast.xml';
-  //   var response = '';
-  //   this.http.get(url, { responseType: 'text' }).subscribe(httpresponse =>
-  //     response=httpresponse
-  //     );
-  //    setTimeout(() => {
-  //     this.jsonRSS = JSON.parse(xml2json(response, {compact: true, spaces: 0, ignoreDeclaration: true, trim: true}));
-  //     console.log(this.jsonRSS.rss.channel.title._text);
-  //     console.log(Object.keys(this.jsonRSS.rss.channel.item).length); //Total Number of elements
-  //   }, 1000)
-
-  //   //Umbau Add Parameter hinzufügen automatich laden und speichern Cover einmalig
-  //   //Umbau Home/Radio Interface und Funktion
-  //   //Media Service
-  // }
-
-  // getRssFeedCover(id: string){
-  //   this.http.get(id, { responseType: 'text' }).subscribe(httpresponse =>
-  //       this.httpResponse=httpresponse);
-  //   setTimeout(() => {
-  //       this.jsonRSS = JSON.parse(xml2json(this.httpResponse, {compact: true, spaces: 0, ignoreDeclaration: true, trim: true}));
-  //       return this.jsonRSS.rss.channel.image.url._text;
-  //   }, 500)
-  // }
 }
