+++ dev/Sonos-Kids-Controller-master/src/app/utils.ts
@@ -1,0 +1,14 @@
+import { Media } from "./media"
+
+export type ExtraDataMedia = Pick<Media, 'artistcover' | 'shuffle' | 'aPartOfAll' | 'aPartOfAllMin' | 'aPartOfAllMax' | 'sorting'>
+
+export namespace Utils {
+    export const copyExtraMediaData = (source: ExtraDataMedia, target: Media): void => {
+        const keys = ['artistcover', 'shuffle',  'aPartOfAll', 'aPartOfAllMin', 'aPartOfAllMax', 'sorting']
+        keys.forEach(key => {
+          if (source[key]) {
+            target[key] = source[key]
+          }
+        })
+      }
+}
